Bundled get_cities, get_companies, get_company in a single test file to only fill the DB once
