+++ tests/functional/test_api.py
@@ -1,6 +1,36 @@
 from flask import json
 
 
+def test_get_cities(client, insert_data_db):
+    '''
+    GIVEN that there are mutiple cities in the DB
+    WHEN the '/api/v1/cities' page is requested (GET)
+    THEN the response will contain a list of all the cities and the count of companies in those cities
+    '''
+    response = client.get('/api/v1/cities')
+
+    data = json.loads(response.get_data(as_text=True))
+
+    assert response.status_code == 200
+    assert len(data) == 11
+    assert data[0][0] == 'Rotterdam'
+
+
+def test_get_cities_city_like(client, init_testdb, insert_data_db):
+    '''
+    GIVEN that Rotterdam and Amsterdam are in the cities DB 
+    WHEN the '/api/v1/cities is requested with a city_like=Dam parameter
+    THEN the response will contain 2 records and their associated data
+    '''
+    response = client.get('/api/v1/cities?city_like=Dam')
+    data = json.loads(response.get_data(as_text=True))
+
+    assert response.status_code == 200
+    assert len(data) == 2
+    assert data[0][0] == 'Rotterdam'
+    assert data[1][0] == 'Amsterdam'
+
+
 def test_get_companies(client, init_testdb, insert_data_db):
     '''
     GIVEN that there are more than 15 companies in the DB
@@ -73,3 +103,18 @@ def test_get_multiple_parameters(client, init_testdb, insert_data_db):
 
     assert response.status_code == 200
     assert len(data['items']) == 1
+
+
+def test_get_company(client, init_testdb, insert_data_db):
+    '''
+    GIVEN a Flask application configured for testing
+    WHEN the '/companies' page is requested (GET) to get a specific company
+    THEN check that the response is valid and particular data can be found
+    '''
+    response = client.get('/api/v1/companies/1')
+
+    data = json.loads(response.get_data(as_text=True))
+
+    assert response.status_code == 200
+    assert data['city_name'] == 'Rotterdam'
+    assert data['company_id'] == 1
